Move EndOfYearStoriesDataSource methods to EodOfYearDataManager
