+++ app/src/main/java/au/com/shiftyjelly/pocketcasts/ui/MainActivityViewModel.kt
@@ -3,10 +3,10 @@ package au.com.shiftyjelly.pocketcasts.ui
 import androidx.lifecycle.LiveData
 import androidx.lifecycle.LiveDataReactiveStreams
 import androidx.lifecycle.ViewModel
-import au.com.shiftyjelly.pocketcasts.endofyear.StoriesDataSource
 import au.com.shiftyjelly.pocketcasts.models.to.SignInState
 import au.com.shiftyjelly.pocketcasts.models.to.SubscriptionStatus
 import au.com.shiftyjelly.pocketcasts.preferences.Settings
+import au.com.shiftyjelly.pocketcasts.repositories.endofyear.EndOfYearManager
 import au.com.shiftyjelly.pocketcasts.repositories.playback.PlaybackManager
 import au.com.shiftyjelly.pocketcasts.repositories.playback.PlaybackState
 import au.com.shiftyjelly.pocketcasts.repositories.user.UserManager
@@ -22,7 +22,7 @@ class MainActivityViewModel
     playbackManager: PlaybackManager,
     userManager: UserManager,
     val settings: Settings,
-    private val storiesDataSource: StoriesDataSource,
+    private val endOfYearManager: EndOfYearManager
 ) : ViewModel() {
 
     var isPlayerOpen: Boolean = false
@@ -51,5 +51,5 @@ class MainActivityViewModel
         return signInState.isExpiredTrial && !settings.getTrialFinishedSeen()
     }
 
-    fun isEndOfYearStoriesEligible() = storiesDataSource.isEligibleForStories()
+    fun isEndOfYearStoriesEligible() = endOfYearManager.isEligibleForStories()
 }
