+++ modules/features/endofyear/src/main/java/au/com/shiftyjelly/pocketcasts/endofyear/StoriesPage.kt
@@ -57,16 +57,6 @@ import au.com.shiftyjelly.pocketcasts.compose.buttons.RowOutlinedButton
 import au.com.shiftyjelly.pocketcasts.compose.components.TextP50
 import au.com.shiftyjelly.pocketcasts.compose.preview.ThemePreviewParameterProvider
 import au.com.shiftyjelly.pocketcasts.endofyear.StoriesViewModel.State
-import au.com.shiftyjelly.pocketcasts.endofyear.stories.Story
-import au.com.shiftyjelly.pocketcasts.endofyear.stories.StoryEpilogue
-import au.com.shiftyjelly.pocketcasts.endofyear.stories.StoryIntro
-import au.com.shiftyjelly.pocketcasts.endofyear.stories.StoryListenedCategories
-import au.com.shiftyjelly.pocketcasts.endofyear.stories.StoryListenedNumbers
-import au.com.shiftyjelly.pocketcasts.endofyear.stories.StoryListeningTime
-import au.com.shiftyjelly.pocketcasts.endofyear.stories.StoryLongestEpisode
-import au.com.shiftyjelly.pocketcasts.endofyear.stories.StoryTopFivePodcasts
-import au.com.shiftyjelly.pocketcasts.endofyear.stories.StoryTopListenedCategories
-import au.com.shiftyjelly.pocketcasts.endofyear.stories.StoryTopPodcast
 import au.com.shiftyjelly.pocketcasts.endofyear.views.convertibleToBitmap
 import au.com.shiftyjelly.pocketcasts.endofyear.views.stories.StoryEpilogueView
 import au.com.shiftyjelly.pocketcasts.endofyear.views.stories.StoryIntroView
@@ -78,6 +68,16 @@ import au.com.shiftyjelly.pocketcasts.endofyear.views.stories.StoryTopFivePodcas
 import au.com.shiftyjelly.pocketcasts.endofyear.views.stories.StoryTopListenedCategoriesView
 import au.com.shiftyjelly.pocketcasts.endofyear.views.stories.StoryTopPodcastView
 import au.com.shiftyjelly.pocketcasts.models.db.helper.ListenedNumbers
+import au.com.shiftyjelly.pocketcasts.repositories.endofyear.stories.Story
+import au.com.shiftyjelly.pocketcasts.repositories.endofyear.stories.StoryEpilogue
+import au.com.shiftyjelly.pocketcasts.repositories.endofyear.stories.StoryIntro
+import au.com.shiftyjelly.pocketcasts.repositories.endofyear.stories.StoryListenedCategories
+import au.com.shiftyjelly.pocketcasts.repositories.endofyear.stories.StoryListenedNumbers
+import au.com.shiftyjelly.pocketcasts.repositories.endofyear.stories.StoryListeningTime
+import au.com.shiftyjelly.pocketcasts.repositories.endofyear.stories.StoryLongestEpisode
+import au.com.shiftyjelly.pocketcasts.repositories.endofyear.stories.StoryTopFivePodcasts
+import au.com.shiftyjelly.pocketcasts.repositories.endofyear.stories.StoryTopListenedCategories
+import au.com.shiftyjelly.pocketcasts.repositories.endofyear.stories.StoryTopPodcast
 import au.com.shiftyjelly.pocketcasts.ui.helper.FragmentHostListener
 import au.com.shiftyjelly.pocketcasts.ui.helper.StatusBarColor
 import au.com.shiftyjelly.pocketcasts.ui.theme.Theme
