--- modules/features/endofyear/src/main/java/au/com/shiftyjelly/pocketcasts/endofyear/di/EndOfYearModule.kt
@@ -1,15 +1,0 @@
-package au.com.shiftyjelly.pocketcasts.endofyear.di
-
-import au.com.shiftyjelly.pocketcasts.endofyear.StoriesDataSource
-import au.com.shiftyjelly.pocketcasts.endofyear.stories.EndOfYearStoriesDataSource
-import dagger.Binds
-import dagger.Module
-import dagger.hilt.InstallIn
-import dagger.hilt.components.SingletonComponent
-
-@Module
-@InstallIn(SingletonComponent::class)
-abstract class EndOfYearModule {
-    @Binds
-    abstract fun providesEndOfYearStoriesDataSource(dataSource: EndOfYearStoriesDataSource): StoriesDataSource
-}
