--- modules/features/endofyear/src/main/java/au/com/shiftyjelly/pocketcasts/endofyear/stories/StoryEpilogue.kt
@@ -1,3 +1,0 @@
-package au.com.shiftyjelly.pocketcasts.endofyear.stories
-
-class StoryEpilogue : Story()
