+++ modules/features/profile/src/main/java/au/com/shiftyjelly/pocketcasts/profile/ProfileViewModel.kt
@@ -4,10 +4,10 @@ import androidx.lifecycle.LiveData
 import androidx.lifecycle.LiveDataReactiveStreams
 import androidx.lifecycle.MutableLiveData
 import androidx.lifecycle.ViewModel
-import au.com.shiftyjelly.pocketcasts.endofyear.StoriesDataSource
 import au.com.shiftyjelly.pocketcasts.models.to.RefreshState
 import au.com.shiftyjelly.pocketcasts.models.to.SignInState
 import au.com.shiftyjelly.pocketcasts.preferences.Settings
+import au.com.shiftyjelly.pocketcasts.repositories.endofyear.EndOfYearManager
 import au.com.shiftyjelly.pocketcasts.repositories.podcast.PodcastManager
 import au.com.shiftyjelly.pocketcasts.repositories.user.StatsManager
 import au.com.shiftyjelly.pocketcasts.repositories.user.UserManager
@@ -21,7 +21,7 @@ class ProfileViewModel @Inject constructor(
     val podcastManager: PodcastManager,
     val statsManager: StatsManager,
     val userManager: UserManager,
-    private val storiesDataSource: StoriesDataSource,
+    private val endOfYearManager: EndOfYearManager
 ) : ViewModel() {
     var isFragmentChangingConfigurations: Boolean = false
     val podcastCount: LiveData<Int> = LiveDataReactiveStreams.fromPublisher(podcastManager.observeCountSubscribed())
@@ -37,7 +37,7 @@ class ProfileViewModel @Inject constructor(
         settings.refreshStateObservable.toFlowable(BackpressureStrategy.LATEST)
     )
 
-    fun isEndOfYearStoriesEligible() = storiesDataSource.isEligibleForStories()
+    fun isEndOfYearStoriesEligible() = endOfYearManager.isEligibleForStories()
 
     fun clearFailedRefresh() {
         val lastSuccess = settings.getLastSuccessRefreshState()
