+++ modules/services/repositories/src/main/java/au/com/shiftyjelly/pocketcasts/repositories/endofyear/EndOfYearManager.kt
@@ -4,9 +4,13 @@ import au.com.shiftyjelly.pocketcasts.models.db.helper.ListenedCategory
 import au.com.shiftyjelly.pocketcasts.models.db.helper.ListenedNumbers
 import au.com.shiftyjelly.pocketcasts.models.db.helper.LongestEpisode
 import au.com.shiftyjelly.pocketcasts.models.db.helper.TopPodcast
+import au.com.shiftyjelly.pocketcasts.repositories.endofyear.stories.Story
 import kotlinx.coroutines.flow.Flow
 
 interface EndOfYearManager {
+    fun isEligibleForStories(): Flow<Boolean>
+    fun hasFullListeningHistory(): Flow<Boolean>
+    fun loadStories(): Flow<List<Story>>
     fun hasEpisodesPlayedUpto(year: Int, playedUpToInSecs: Long): Flow<Boolean>
     fun hasEpisodeInteractedBefore(year: Int): Flow<Boolean>
     fun hasListeningHistoryEpisodesInLimitForYear(year: Int, limit: Int): Flow<Boolean>
