+++ modules/services/repositories/src/main/java/au/com/shiftyjelly/pocketcasts/repositories/endofyear/EndOfYearManagerImpl.kt
@@ -1,17 +1,30 @@
 package au.com.shiftyjelly.pocketcasts.repositories.endofyear
 
 import au.com.shiftyjelly.pocketcasts.models.db.helper.ListenedNumbers
+import au.com.shiftyjelly.pocketcasts.repositories.BuildConfig
+import au.com.shiftyjelly.pocketcasts.repositories.endofyear.stories.Story
+import au.com.shiftyjelly.pocketcasts.repositories.endofyear.stories.StoryEpilogue
+import au.com.shiftyjelly.pocketcasts.repositories.endofyear.stories.StoryIntro
+import au.com.shiftyjelly.pocketcasts.repositories.endofyear.stories.StoryListenedCategories
+import au.com.shiftyjelly.pocketcasts.repositories.endofyear.stories.StoryListenedNumbers
+import au.com.shiftyjelly.pocketcasts.repositories.endofyear.stories.StoryListeningTime
+import au.com.shiftyjelly.pocketcasts.repositories.endofyear.stories.StoryLongestEpisode
+import au.com.shiftyjelly.pocketcasts.repositories.endofyear.stories.StoryTopFivePodcasts
+import au.com.shiftyjelly.pocketcasts.repositories.endofyear.stories.StoryTopListenedCategories
+import au.com.shiftyjelly.pocketcasts.repositories.endofyear.stories.StoryTopPodcast
 import au.com.shiftyjelly.pocketcasts.repositories.podcast.EpisodeManager
 import au.com.shiftyjelly.pocketcasts.repositories.podcast.PodcastManager
 import au.com.shiftyjelly.pocketcasts.utils.DateUtil
 import kotlinx.coroutines.CoroutineScope
 import kotlinx.coroutines.Dispatchers
 import kotlinx.coroutines.flow.Flow
+import kotlinx.coroutines.flow.combine
 import kotlinx.coroutines.flow.flowOf
 import kotlinx.coroutines.flow.transform
 import timber.log.Timber
 import java.time.DateTimeException
 import java.time.LocalDate
+import java.util.concurrent.TimeUnit
 import javax.inject.Inject
 import kotlin.coroutines.CoroutineContext
 
@@ -21,6 +34,58 @@ class EndOfYearManagerImpl @Inject constructor(
 ) : EndOfYearManager, CoroutineScope {
     override val coroutineContext: CoroutineContext
         get() = Dispatchers.Default
+
+    override fun isEligibleForStories(): Flow<Boolean> =
+        hasEpisodesPlayedUpto(YEAR, TimeUnit.MINUTES.toSeconds(EPISODE_MINIMUM_PLAYED_TIME_IN_MIN))
+            .transform { emit(it && BuildConfig.END_OF_YEAR_ENABLED) }
+
+    /* Check if the user has the full listening history or not.
+     This is not 100% accurate. In order to determine if the user
+     has the full history we check for their latest episode listened.
+     If this episode was interacted in 2021 or before, we assume they
+     have the full history.
+     If this is not true, we check for the total number of items of
+     this year. If the number is less than or equal 100, we assume they
+     have the full history. */
+    override fun hasFullListeningHistory(): Flow<Boolean> =
+        combine(
+            hasEpisodeInteractedBefore(YEAR),
+            hasListeningHistoryEpisodesInLimitForYear(YEAR, limit = LISTENING_HISTORY_LIMIT),
+        ) { hasEpisodeInteractedBefore, hasListeningHistoryEpisodesInLimitForYear ->
+            hasEpisodeInteractedBefore || hasListeningHistoryEpisodesInLimitForYear
+        }
+
+    override fun loadStories(): Flow<List<Story>> {
+        return combine(
+            getTotalListeningTimeInSecsForYear(YEAR),
+            findListenedCategoriesForYear(YEAR),
+            findListenedNumbersForYear(YEAR),
+            findTopPodcastsForYear(YEAR, limit = 10),
+            findLongestPlayedEpisodeForYear(YEAR)
+        ) { listeningTime, listenedCategories, listenedNumbers, topPodcasts, longestEpisode ->
+            val stories = mutableListOf<Story>()
+
+            stories.add(StoryIntro())
+            listeningTime?.let { stories.add(StoryListeningTime(it, topPodcasts.takeLast(3))) }
+            if (listenedCategories.isNotEmpty()) {
+                stories.add(StoryListenedCategories(listenedCategories))
+                stories.add(StoryTopListenedCategories(listenedCategories))
+            }
+            if (listenedNumbers.numberOfEpisodes > 1 && listenedNumbers.numberOfPodcasts > 1) {
+                stories.add(StoryListenedNumbers(listenedNumbers))
+            }
+            if (topPodcasts.isNotEmpty()) {
+                stories.add(StoryTopPodcast(topPodcasts.first()))
+                if (topPodcasts.size > 1) {
+                    stories.add(StoryTopFivePodcasts(topPodcasts.take(5)))
+                }
+            }
+            longestEpisode?.let { stories.add(StoryLongestEpisode(it)) }
+            stories.add(StoryEpilogue())
+
+            stories
+        }
+    }
 
     /* Returns whether user listened to at least one episode for more than given time for the year */
     override fun hasEpisodesPlayedUpto(year: Int, playedUpToInSecs: Long): Flow<Boolean> =
@@ -81,4 +146,10 @@ class EndOfYearManagerImpl @Inject constructor(
     }
 
     data class YearStartAndEndEpochMs(val start: Long, val end: Long)
+
+    companion object {
+        private const val YEAR = 2022
+        private const val EPISODE_MINIMUM_PLAYED_TIME_IN_MIN = 30L
+        private const val LISTENING_HISTORY_LIMIT = 100
+    }
 }
