+++ modules/services/repositories/src/main/java/au/com/shiftyjelly/pocketcasts/repositories/endofyear/stories/StoryEpilogue.kt
@@ -1,0 +1,3 @@
+package au.com.shiftyjelly.pocketcasts.repositories.endofyear.stories
+
+class StoryEpilogue : Story()
