+++ modules/services/repositories/src/main/java/au/com/shiftyjelly/pocketcasts/repositories/endofyear/stories/StoryIntro.kt
@@ -1,4 +1,4 @@
-package au.com.shiftyjelly.pocketcasts.endofyear.stories
+package au.com.shiftyjelly.pocketcasts.repositories.endofyear.stories
 
 import androidx.compose.ui.graphics.Color
 class StoryIntro : Story() {
