+++ modules/services/repositories/src/main/java/au/com/shiftyjelly/pocketcasts/repositories/endofyear/stories/StoryListenedCategories.kt
@@ -1,7 +1,7 @@
-package au.com.shiftyjelly.pocketcasts.endofyear.stories
+package au.com.shiftyjelly.pocketcasts.repositories.endofyear.stories
 
 import au.com.shiftyjelly.pocketcasts.models.db.helper.ListenedCategory
 
-class StoryTopListenedCategories(
+class StoryListenedCategories(
     val listenedCategories: List<ListenedCategory>
 ) : Story()
