+++ modules/services/repositories/src/main/java/au/com/shiftyjelly/pocketcasts/repositories/endofyear/stories/StoryTopFivePodcasts.kt
@@ -1,4 +1,4 @@
-package au.com.shiftyjelly.pocketcasts.endofyear.stories
+package au.com.shiftyjelly.pocketcasts.repositories.endofyear.stories
 
 import au.com.shiftyjelly.pocketcasts.models.db.helper.TopPodcast
